+++ Curtains/Curtains.Application/CurtainsService/BedspreadsService.cs
@@ -4,6 +4,8 @@ using Curtains.Application.DTO;
 using Curtains.Domain.Models;
 using Curtains.Infrastructure.Interfaces;
 using Curtains.Infrastructure.Repositories;
+using Curtains.Infrastructure.Shared.Exceptions;
+using Microsoft.Extensions.Logging;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -19,11 +21,13 @@ namespace Curtains.Application.CurtainsService
     {
         private readonly IBedspreadsRepository _bedspreadsRepository;
         private readonly IMapper _mapper;
+        private readonly ILogger _logger;
 
-        public BedspreadsService(IBedspreadsRepository bedspreadsRepository, IMapper mapper)
+        public BedspreadsService(IBedspreadsRepository bedspreadsRepository, IMapper mapper, ILogger logger)
         {
             _bedspreadsRepository = bedspreadsRepository;
             _mapper = mapper;
+            _logger = logger;
         }
 
         public IEnumerable<BedspreadsDTO> GetAll()
@@ -35,6 +39,11 @@ namespace Curtains.Application.CurtainsService
         public async Task<BedspreadsDTO> GetByIdAsync(int Id)
         {
             var bedspreads = await _bedspreadsRepository.GetByIdAsync(Id);
+            if (bedspreads == null)
+            {
+                _logger.LogError("Bedspreads model is null");
+                throw new ResourceNotFoundException("Bedspreads model is null");
+            }
             var bedspreadsDTO = _mapper.Map<BedspreadsDTO>(bedspreads);
             return bedspreadsDTO;
         }
@@ -42,18 +51,33 @@ namespace Curtains.Application.CurtainsService
         public async Task InsertAsync(BedspreadsDTO entity, CancellationToken cancelationToken)
         {
             var bedspread = _mapper.Map<BedspreadsModel>(entity);
+            if (bedspread == null)
+            {
+                _logger.LogError("Bedspreads model is null");
+                throw new ResourceNotFoundException("Bedspreads model is null");
+            }
             await _bedspreadsRepository.InsertAsync(bedspread, cancelationToken);
         }
 
         public async Task RemoveAsync(BedspreadsDTO entity)
         {
             var bedspread = _mapper.Map<BedspreadsModel>(entity);
+            if (bedspread == null)
+            {
+                _logger.LogError("Bedspreads model is null");
+                throw new ResourceNotFoundException("Bedspreads model is null");
+            }
             await _bedspreadsRepository.RemoveAsync(bedspread);
         }
 
         public async Task UpdateAsync(BedspreadsDTO entity)
         {
             var bedspread = _mapper.Map<BedspreadsModel>(entity);
+            if (bedspread == null)
+            {
+                _logger.LogError("Bedspreads model is null");
+                throw new ResourceNotFoundException("Bedspreads model is null");
+            }
             await _bedspreadsRepository.UpdateAsync(bedspread);
         }
     }
