+++ Curtains/Curtains.Application/CurtainsService/BracingService.cs
@@ -4,8 +4,11 @@ using Curtains.Application.DTO;
 using Curtains.Domain.Models;
 using Curtains.Infrastructure.Interfaces;
 using Curtains.Infrastructure.Repositories;
+using Curtains.Infrastructure.Shared.Exceptions;
+using Microsoft.Extensions.Logging;
 using System;
 using System.Collections.Generic;
+using System.Drawing;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -19,11 +22,13 @@ namespace Curtains.Application.CurtainsService
     {
         private readonly IBracingRepository _bracingRepository;
         private readonly IMapper _mapper;
+        private readonly ILogger _logger;
 
-        public BracingService(IBracingRepository bracingRepository, IMapper mapper)
+        public BracingService(IBracingRepository bracingRepository, IMapper mapper, ILogger logger)
         {
             _bracingRepository = bracingRepository;
             _mapper = mapper;
+            _logger = logger;
         }
 
         public IEnumerable<BracingDTO> GetAll()
@@ -35,6 +40,11 @@ namespace Curtains.Application.CurtainsService
         public async Task<BracingDTO> GetByIdAsync(int Id)
         {
             var bracing = await _bracingRepository.GetByIdAsync(Id);
+            if (bracing == null)
+            {
+                _logger.LogError("bracing model is null");
+                throw new ResourceNotFoundException("bracing model is null");
+            }
             var bracingDTO = _mapper.Map<BracingDTO>(bracing);
             return bracingDTO;
         }
@@ -42,18 +52,33 @@ namespace Curtains.Application.CurtainsService
         public async Task InsertAsync(BracingDTO entity, CancellationToken cancelationToken)
         {
             var bracing = _mapper.Map<BracingModel>(entity);
+            if (bracing == null)
+            {
+                _logger.LogError("bracing model is null");
+                throw new ResourceNotFoundException("bracing model is null");
+            }
             await _bracingRepository.InsertAsync(bracing, cancelationToken);
         }
 
         public async Task RemoveAsync(BracingDTO entity)
         {
             var bracing = _mapper.Map<BracingModel>(entity);
+            if (bracing == null)
+            {
+                _logger.LogError("bracing model is null");
+                throw new ResourceNotFoundException("bracing model is null");
+            }
             await _bracingRepository.RemoveAsync(bracing);
         }
 
         public async Task UpdateAsync(BracingDTO entity)
         {
             var bracing = _mapper.Map<BracingModel>(entity);
+            if (bracing == null)
+            {
+                _logger.LogError("bracing model is null");
+                throw new ResourceNotFoundException("bracing model is null");
+            }
             await _bracingRepository.UpdateAsync(bracing);
         }
     }
