+++ Curtains/Curtains.Application/CurtainsService/CurtainsService.cs
@@ -4,6 +4,8 @@ using Curtains.Application.DTO;
 using Curtains.Domain.Models;
 using Curtains.Infrastructure.Interfaces;
 using Curtains.Infrastructure.Repositories;
+using Curtains.Infrastructure.Shared.Exceptions;
+using Microsoft.Extensions.Logging;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -19,11 +21,13 @@ namespace Curtains.Application.CurtainsService
     {
         private readonly ICurtainsRepository _curtainsRepository;
         private readonly IMapper _mapper;
+        private readonly ILogger _logger;
 
-        public CurtainsService(ICurtainsRepository curtainsRepository, IMapper mapper)
+        public CurtainsService(ICurtainsRepository curtainsRepository, IMapper mapper, ILogger logger)
         {
             _curtainsRepository = curtainsRepository;
             _mapper = mapper;
+            _logger = logger;
         }
 
         public IEnumerable<CurtainsDTO> GetAll()
@@ -35,6 +39,11 @@ namespace Curtains.Application.CurtainsService
         public async Task<CurtainsDTO> GetByIdAsync(int Id)
         {
             var curtains = await _curtainsRepository.GetByIdAsync(Id);
+            if (curtains == null)
+            {
+                _logger.LogError("Сurtians model is null");
+                throw new ResourceNotFoundException("Сurtians model is null");
+            }
             var curtainsDTO = _mapper.Map<CurtainsDTO>(curtains);
             return curtainsDTO;
         }
@@ -42,18 +51,33 @@ namespace Curtains.Application.CurtainsService
         public async Task InsertAsync(CurtainsDTO entity, CancellationToken cancelationToken)
         {
             var curtains = _mapper.Map<CurtainsModel>(entity);
+            if (curtains == null)
+            {
+                _logger.LogError("Сurtians model is null");
+                throw new ResourceNotFoundException("Сurtians model is null");
+            }
             await _curtainsRepository.InsertAsync(curtains, cancelationToken);
         }
 
         public async Task RemoveAsync(CurtainsDTO entity)
         {
             var curtains = _mapper.Map<CurtainsModel>(entity);
+            if (curtains == null)
+            {
+                _logger.LogError("Сurtians model is null");
+                throw new ResourceNotFoundException("Сurtians model is null");
+            }
             await _curtainsRepository.RemoveAsync(curtains);
         }
 
         public async Task UpdateAsync(CurtainsDTO entity)
         {
             var curtains = _mapper.Map<CurtainsModel>(entity);
+            if (curtains == null)
+            {
+                _logger.LogError("Сurtians model is null");
+                throw new ResourceNotFoundException("Сurtians model is null");
+            }
             await _curtainsRepository.UpdateAsync(curtains);
         }
     }
