+++ Curtains/Curtains.Application/CurtainsService/ProductSetService.cs
@@ -4,6 +4,8 @@ using Curtains.Application.DTO;
 using Curtains.Domain.Models;
 using Curtains.Infrastructure.Interfaces;
 using Curtains.Infrastructure.Repositories;
+using Curtains.Infrastructure.Shared.Exceptions;
+using Microsoft.Extensions.Logging;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -19,11 +21,13 @@ namespace Curtains.Application.CurtainsService
     {
         private readonly IProductSetRepository _productSetRepository;
         private readonly IMapper _mapper;
+        private readonly ILogger _logger;
 
-        public ProductSetService(IProductSetRepository productSetRepository, IMapper mapper)
+        public ProductSetService(IProductSetRepository productSetRepository, IMapper mapper, ILogger logger)
         {
             _productSetRepository = productSetRepository;
             _mapper = mapper;
+            _logger = logger;
         }
 
         public IEnumerable<ProductSetDTO> GetAll()
@@ -35,6 +39,11 @@ namespace Curtains.Application.CurtainsService
         public async Task<ProductSetDTO> GetByIdAsync(int Id)
         {
             var productSet = await _productSetRepository.GetByIdAsync(Id);
+            if (productSet == null)
+            {
+                _logger.LogError("Product set model is null");
+                throw new ResourceNotFoundException("Product se is null");
+            }
             var productSetDTO = _mapper.Map<ProductSetDTO>(productSet);
             return productSetDTO;
         }
@@ -42,18 +51,33 @@ namespace Curtains.Application.CurtainsService
         public async Task InsertAsync(ProductSetDTO entity, CancellationToken cancelationToken)
         {
             var productSet = _mapper.Map<ProductSetModel>(entity);
+            if (productSet == null)
+            {
+                _logger.LogError("Product set model is null");
+                throw new ResourceNotFoundException("Product se is null");
+            }
             await _productSetRepository.InsertAsync(productSet, cancelationToken);
         }
 
         public async Task RemoveAsync(ProductSetDTO entity)
         {
             var productSet = _mapper.Map<ProductSetModel>(entity);
+            if (productSet == null)
+            {
+                _logger.LogError("Product set model is null");
+                throw new ResourceNotFoundException("Product se is null");
+            }
             await _productSetRepository.RemoveAsync(productSet);
         }
 
         public async Task UpdateAsync(ProductSetDTO entity)
         {
             var productSet = _mapper.Map<ProductSetModel>(entity);
+            if (productSet == null)
+            {
+                _logger.LogError("Product set model is null");
+                throw new ResourceNotFoundException("Product se is null");
+            }
             await _productSetRepository.UpdateAsync(productSet);
         }
     }
